fixed unsigned comparison warning

+++ src/leveldb/util/logging.cc
@@ -54,7 +54,7 @@ bool ConsumeDecimalNumber(Slice* in, uint64_t* val) {
       ++digits;
       const unsigned int delta = (c - '0');
       static const uint64_t kMaxUint64 = ~static_cast<uint64_t>(0);
-      if (v > kMaxUint64/10 || (v == kMaxUint64/10 && delta > kMaxUint64%10)) {
+      if (v > kMaxUint64/10 || (v == (unsigned int)kMaxUint64/10 && delta > (unsigned int)kMaxUint64%10)) {
         // Overflow
         return false;
       }
